[pro] Implement progress color on UIA panel overlays
If a switch or something on the UIA must be flipped in the correct direction (or turned on/off), then that switch will be highlighted in red. The switch will become green once the user completes the task (e.g. flips switch in correct direction).

For tasks that require waiting (O2 and H2O tanks), the switches will be red, then stay yellow while the tanks fill/deplete, then turn green once complete.
Lastly, some logic was updated for when to spawn overlays (some of it was wrong between steps 5 and 6 before).

For future reference, the colors for the overlays correspond to the colors used in the task progress text.

Also, these are the overlay array indexes:
Overlay array indexes
0 -> power ev1
1 -> power ev2
2 -> o2 ev1
3 -> o2 ev2
4 -> o2 vent
5 -> h2o waste ev1
6 -> h2o waste ev2