[tel+pro] Implement new EVTDisplay, fix Ingress progress
Implemented the new EVTDisplay design into HMD scene.
There was a bug with the progress bars on ingress not working properly. I updated the fill logic and it works fine now.
Also, there was a bug when the user would switch from egress to ingress, specifically when switching from complete egress to new ingress. Both ingress and egress were using the same step counter, so there was an out of bounds error happening, and the progress of both egress/ingress was broken. Fixed by clearing progress between modes.